Rename fraud proof example to game256
